django-apis/requirements.txt
asgiref==3.4.1
auditlog3==1.0.1
autopep8==1.7.0
cachetools==5.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Deprecated==1.2.12
dj-database-url==1.0.0
dj3-cloudinary-storage==0.0.6
Django==3.2.5
django-admin-tools==0.9.2
django-admin-tools-stats==0.9.0
django-autoslug==1.9.8
django-bower==5.2.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-nested-admin==3.4.0
django-nvd3==0.9.7
django-oauth-toolkit==1.5.0
django-qsstats-magic==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djcacheutils==3.0.0
drf-yasg==1.20.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.21.0
idna==3.2
inflection==0.5.1
itemadapter==0.7.0
itypes==1.2.0
Jinja2==3.0.1
json5==0.9.10
jwcrypto==0.9.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.1
protobuf==4.21.9
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.20.0
python-memcached==1.59
python-monkey-business==1.0.0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2021.1
requests==2.26.0
rsa==4.9
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
stripe==4.2.0
tld==0.12.6
toml==0.10.2
Unidecode==1.3.6
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==6.2.0
wrapt==1.12.1
